CI(macos): only copy Qt libs if they don't exist
This is mainly a fix for users that try to build obs-websocket themselves.
From a previous (failed) attempt, the Qt libraries may already exist in
the build directory, however marked as read-only. Therefore, cp cannot
overwrite them, even though it would be perfectly fine to do so.
This commit makes it so the copy is not attempted if the libraries exist.

+++ CI/package-macos.sh
@@ -21,8 +21,8 @@ export FILENAME="obs-websocket-$VERSION.pkg"
 export LATEST_FILENAME="obs-websocket-latest-$LATEST_VERSION.pkg"
 
 echo "-- Copying Qt dependencies"
-cp $WS_LIB ./build
-cp $NET_LIB ./build
+if [ ! -f ./build/$(basename $WS_LIB) ]; then cp $WS_LIB ./build; fi
+if [ ! -f ./build/$(basename $NET_LIB) ]; then cp $NET_LIB ./build; fi
 
 chmod +rw ./build/QtWebSockets ./build/QtNetwork
